+++ blocks/body-builder-news/body-builder-news.js
@@ -1,6 +1,7 @@
 import {
-  createOptimizedPicture, getMetadata, getOrigin, readBlockConfig,
-} from '../../scripts/lib-franklin.js';
+  createOptimizedPicture, getMetadata, readBlockConfig,
+} from '../../scripts/aem.js';
+import { getOrigin } from '../../scripts/scripts.js';
 import { ffetch } from '../../scripts/lib-ffetch.js';
 
 const locale = getMetadata('locale');
