+++ blocks/get-an-offer/get-an-offer.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import {
   decorateSections, decorateBlocks, loadBlocks, decorateButtons, decorateIcons, loadCSS,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
 
 export async function showOffer(a) {
   const { href, title } = a;
