+++ blocks/header/header.js
@@ -4,7 +4,7 @@ import {
   getTextLabel,
   getLanguagePath,
 } from '../../scripts/common.js';
-import { createOptimizedPicture, decorateIcons, getMetadata } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture, decorateIcons, getMetadata } from '../../scripts/aem.js';
 
 const blockClass = 'header';
 const disableSearch = getMetadata('disable-search').toLowerCase() === 'true';
