+++ blocks/key-facts/key-facts.css
@@ -14,8 +14,7 @@
   justify-content: center;
   padding: 0;
   flex-flow: row wrap;
-  row-gap: 32px;
-  column-gap: 8px;
+  gap: 32px 8px;
   text-align: center;
 }
