+++ blocks/press-releases/press-releases.js
@@ -7,10 +7,12 @@ import {
 import {
   createOptimizedPicture,
   getMetadata,
-  getOrigin,
   readBlockConfig,
   toClassName,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
+import {
+  getOrigin,
+} from '../../scripts/scripts.js';
 
 const locale = getMetadata('locale');
 const stopWords = ['a', 'an', 'the', 'and', 'to', 'for', 'i', 'of', 'on', 'into'];
