+++ blocks/teaser-grid/teaser-grid.css
@@ -51,8 +51,7 @@
     margin: 0;
     padding: 0;
     display: grid;
-    row-gap: var(--grid-row-gap);
-    column-gap: var(--grid-col-gap);
+    gap: var(--grid-row-gap) var(--grid-col-gap);
     grid-template: repeat(calc(var(--grid-col-count) * var(--grid-row-count)), 1fr) / 100%;
 }
