+++ blocks/v2-content-carousel/v2-content-carousel.js
@@ -1,5 +1,5 @@
 import { createElement, adjustPretitle } from '../../scripts/common.js';
-import { decorateIcons } from '../../scripts/lib-franklin.js';
+import { decorateIcons } from '../../scripts/aem.js';
 import { smoothScrollHorizontal } from '../../scripts/motion-helper.js';
 
 const blockName = 'v2-content-carousel';
