+++ blocks/v2-event-notify/v2-event-notify.js
@@ -1,6 +1,6 @@
 import {
   loadBlock, sampleRUM,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
 import {
   createElement,
 } from '../../scripts/common.js';
