+++ blocks/v2-masonry-grid/v2-masonry-grid.js
@@ -1,4 +1,4 @@
-import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture } from '../../scripts/aem.js';
 import {
   isVideoLink,
   createVideo,
