+++ blocks/v2-newsletter/v2-newsletter.js
@@ -1,6 +1,6 @@
 import {
   loadBlock, sampleRUM,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
 import { getTextLabel, createElement } from '../../scripts/common.js';
 
 const blockName = 'v2-newsletter';
