+++ blocks/v2-truck-lineup/v2-truck-lineup.js
@@ -261,7 +261,9 @@ export default function decorate(block) {
   // Update text position + navigation line when page is resized
   window.addEventListener('resize', () => {
     const activeItem = imagesContainer.querySelector(`.${blockName}__image-item.active`);
-    const index = [...activeItem.parentNode.children].indexOf(activeItem);
-    updateActiveItem(index);
+    if (activeItem) {
+      const index = [...activeItem.parentNode.children].indexOf(activeItem);
+      updateActiveItem(index);
+    }
   });
 }
