+++ head.html
@@ -1,6 +1,6 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
 <meta name="referrer" content="no-referrer-when-downgrade" />
-<script src="/scripts/lib-franklin.js" type="module"></script>
+<script src="/scripts/aem.js" type="module"></script>
 <script src="/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/styles/styles.css"/>
 <link rel="icon" href="data:,">
