+++ package.json
@@ -1,12 +1,10 @@
 {
   "name": "@adobe/helix-project-boilerplate",
   "private": true,
-  "version": "1.0.0",
+  "version": "1.3.0",
   "description": "Starter project for Adobe Helix",
   "scripts": {
     "dev": "npx @adobe/helix-cli up",
-    "test": "wtr \"./test/**/*.test.js\" --node-resolve --port=2000 --coverage",
-    "test:watch": "npm test -- --watch",
     "lint:js": "eslint .",
     "lint:css": "stylelint 'blocks/**/*.css' 'styles/*.css' 'templates/**/*.css'",
     "lint": "npm run lint:js && npm run lint:css",
@@ -27,19 +25,19 @@
   },
   "homepage": "https://github.com/adobe/helix-project-boilerplate#readme",
   "devDependencies": {
-    "@babel/core": "7.22.10",
-    "@babel/eslint-parser": "7.22.10",
+    "@babel/core": "7.24.4",
+    "@babel/eslint-parser": "7.24.1",
     "@esm-bundle/chai": "4.3.4-fix.0",
     "@web/test-runner": "0.15.0",
     "@web/test-runner-commands": "0.6.5",
     "chai": "4.3.7",
-    "eslint": "8.47.0",
+    "eslint": "8.57.0",
     "eslint-config-airbnb-base": "15.0.0",
-    "eslint-plugin-import": "2.28.0",
+    "eslint-plugin-import": "2.29.1",
     "husky": "^8.0.3",
     "lint-staged": "^14.0.0",
     "sinon": "15.0.1",
-    "stylelint": "15.10.2",
-    "stylelint-config-standard": "34.0.0"
+    "stylelint": "16.3.1",
+    "stylelint-config-standard": "36.0.0"
   }
 }
