--- test/scripts/body.html
@@ -1,59 +1,0 @@
-<header></header>
-<main>
-    <div>
-        <p>
-            <picture>
-                <source media="(max-width: 400px)" srcset="./media_mock.png?width=750&amp;format=webply&amp;optimize=medium">
-                <img src="/test/scripts/media_mock.png?width=2000&amp;format=webply&amp;optimize=medium" alt="" loading="eager">
-            </picture>
-        </p>
-        <h1 id="this-is-a-heading-1">This is a Heading 1</h1>
-        <h2 id="this-is-a-heading-2">This is a Heading 2</h2>
-        <h3 id="this-is-a-heading-3">This is a Heading 3</h3>
-        <h4 id="this-is-a-heading-4">This is a Heading 4</h4>
-        <h5 id="this-is-a-heading-5">This is a Heading 5</h5>
-        <h6 id="this-is-a-heading-6">This is a Heading 6</h6>
-        <ul>
-            <li>Red</li>
-            <li>Green</li>
-            <li>Blue</li>
-        </ul>
-        <ol>
-            <li>First</li>
-            <li>Second</li>
-            <li>Third</li>
-        </ol>
-        <p>Lorem ipsum dolor sit amet, <strong>consectetur</strong> adipiscing elit, sed do eiusmod tempor incididunt ut
-            labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-            aliquip ex ea commodo <em>consequat</em>. Duis aute irure dolor in reprehenderit in voluptate velit esse
-            cillum dolore eu fugiat <code>nulla</code> pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-            culpa qui officia deserunt mollit anim id est <a href="https://www.hlx.live/">laborum</a>.</p>
-        <p>
-            <picture>
-                <source media="(max-width: 400px)"
-                    srcset="/test/scripts/media_mock.png?width=750&amp;format=webply&amp;optimize=medium">
-                <img src="/test/scripts/media_mock.png?width=2000&amp;format=webply&amp;optimize=medium"
-                    alt="" loading="eager">
-            </picture>
-        </p>
-        <p><a href="https://www.hlx.live/">Button</a></p>
-        <p><strong><a href="https://www.hlx.live/">Button Bold</a></strong></p>
-        <p><em><a href="https://www.hlx.live/">Button Italic</a></em></p>
-        <p><strong><em><a href="https://www.hlx.live/">Button Italic Bold</a></em></strong></p>
-        <h3 id="testing-a-code-block">Testing a code block</h3>
-        <pre><code>&lt;meta name="viewport" content="width=device-width, initial-scale=1"/&gt;
-&lt;script src="/scripts.js" type="module"&gt;&lt;/script&gt;
-&lt;link rel="stylesheet" href="/styles.css"/&gt;
-        </code></pre>
-        <div class="hero">
-            <p>Lorem ipsum dolor sit amet, <strong>consectetur</strong> adipiscing elit, sed do eiusmod tempor incididunt ut
-                labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-                aliquip ex ea commodo <em>consequat</em>. Duis aute irure dolor in reprehenderit in voluptate velit esse
-                cillum dolore eu fugiat <code>nulla</code> pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-                culpa qui officia deserunt mollit anim id est <a href="https://www.hlx.live/">laborum</a>.</p>
-        </div>
-        <div></div><!-- test block without name -->
-    </div>
-    <div></div><!-- test removal of empty sections -->
-</main>
-<footer></footer>
