--- test/scripts/config.html
@@ -1,25 +1,0 @@
-<div class="config">
-    <div>
-        <div>Prop 0</div>
-        <div>Plain text</div>
-    </div>
-    <div>
-        <div>Prop 1</div>
-        <div><p>Paragraph</p></div>
-    </div>
-    <div>
-        <div>Prop 2</div>
-        <div><p>First paragraph</p><p>Second paragraph</p></div>
-    </div>
-    <div>
-        <div>Prop 3</div>
-        <div><a href="https://www.adobe.com/">Link</a></div>
-    </div>
-    <div>
-        <div>Prop 4</div>
-        <div>
-            <a href="https://www.adobe.com/">First link</a>
-            <a href="https://www.hlx.live/">Second link</a>
-        </div>
-    </div>
-</div>
