--- test/scripts/dummy.html
@@ -1,4 +1,0 @@
-<!-- dummy html to run scripts.js against in a first pass -->
-<header></header>
-<main></main>
-<footer></footer>
