--- test/scripts/head.html
@@ -1,12 +1,0 @@
-<title>Foo</title>
-<link rel="canonical" href="https://foo.bar/">
-<meta name="description" content="Lorem ipsum dolor sit amet.">
-<meta property="og:title" content="Foo">
-<meta property="og:description" content="Lorem ipsum dolor sit amet.">
-<meta property="og:url" content="https://foo.bar/">
-<meta property="og:image" content="https://foo.bar/media_12637fbb67cddc5d293b20975e89028d919270ac0.jpeg?width=1200&amp;format=pjpg&amp;optimize=medium">
-<meta property="og:image:secure_url" content="https://foo.bar/media_12637fbb67cddc5d293b20975e89028d919270ac0.jpeg?width=1200&amp;format=pjpg&amp;optimize=medium"">
-<meta name="twitter:title" content="Foo">
-<meta name="twitter:description" content="Lorem ipsum dolor sit amet.">
-<meta name="twitter:image" content="https://foo.bar/media_12637fbb67cddc5d293b20975e89028d919270ac0.jpeg?width=1200&amp;format=pjpg&amp;optimize=medium">
-<meta name="viewport" content="width=device-width, initial-scale=1">
