--- test/scripts/scripts.test.js
@@ -1,43 +1,0 @@
-/* eslint-disable no-unused-expressions */
-/* global describe before it */
-
-import { readFile } from '@web/test-runner-commands';
-import { expect } from '@esm-bundle/chai';
-import sinon from 'sinon';
-
-/** @type {import('./types').Scripts} */
-// eslint-disable-next-line no-unused-vars
-let scripts;
-/** @type {import('./types').LibFranklin} */
-let lib;
-
-document.body.innerHTML = await readFile({ path: './dummy.html' });
-document.head.innerHTML = await readFile({ path: './head.html' });
-
-describe('Core Helix features', () => {
-  before(async () => {
-    scripts = await import('../../scripts/scripts.js');
-    lib = await import('../../scripts/lib-franklin.js');
-
-    document.body.innerHTML = await readFile({ path: './body.html' });
-  });
-
-  it('Initializes window.hlx', async () => {
-    // simulate code base path and turn on lighthouse
-    document.head.appendChild(document.createElement('script')).src = '/foo/scripts/scripts.js';
-    window.history.pushState({}, '', `${window.location.href}&lighthouse=on`);
-    lib.setup();
-
-    expect(window.hlx.codeBasePath).to.equal('/foo');
-    expect(window.hlx.lighthouse).to.equal(true);
-
-    // test error handling
-    const url = sinon.stub(window, 'URL');
-
-    // cleanup
-    url.restore();
-    window.hlx.codeBasePath = '';
-    window.hlx.lighthouse = false;
-    Array.from(document.querySelectorAll('script')).pop().remove();
-  });
-});
