--- test/scripts/types.d.ts
@@ -1,5 +1,0 @@
-import type * as ScriptsModule from '../../scripts/scripts';
-import type * as LibModule from '../../scripts/lib-franklin';
-
-export type Scripts = typeof ScriptsModule;
-export type LibFranklin = typeof LibModule;
